ajout exercice java

+++ Algo/Java/mon1erProjet/src/exercice1/App.java
@@ -21,18 +21,18 @@ FIN DU PROGRAMME
 		
 		//Declaration des variable
 		
-		double nombre1;
-		double nombre2;
+		int nombre1;
+		int nombre2;
 		double moyenne;
 		
 		Scanner sc = new Scanner(System.in);
 		
 		System.out.println("saisir le 1er nombre");
-		nombre1 = sc.nextDouble();
+		nombre1 = sc.nextInt();
 		System.out.println("saisir le 2eme nombre");
-		nombre2 = sc.nextDouble();
+		nombre2 = sc.nextInt();
 
-		moyenne = (nombre1 + nombre2) / 2;
+		moyenne = (nombre1 + nombre2) / 2d;
 		
 		System.out.println("la moyenne de "+nombre1+" et "+nombre2+" est "+moyenne);
